[test][compiler] Fix compilation error on Android arm64 bot
Fix for
https://ci.chromium.org/ui/p/v8/builders/ci/V8%20Android%20Arm64%20-%20builder/56497/overview

No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5244647
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#92052}

+++ test/unittests/compiler/turboshaft/reducer-test.h
@@ -168,7 +168,7 @@ class TestInstance {
       stream_ = std::make_unique<std::ofstream>(file_name.str(),
                                                 std::ios_base::trunc);
       *stream_ << "{\"function\" : ";
-      int len = strlen("test_generated_function") + 1;
+      size_t len = strlen("test_generated_function") + 1;
       auto name = std::make_unique<char[]>(len);
       snprintf(name.get(), len, "test_generated_function");
       JsonPrintFunctionSource(*stream_, -1, std::move(name), Handle<Script>{},
